P1/LS/requirements.txt
# Django Core
Django==5.1.6

# Django REST Framework para API
djangorestframework==3.14.0

# HTTP Requests para APIs
requests==2.31.0

# CORS headers para API
django-cors-headers==4.3.1

# Filtrado avanzado para API
django-filter==23.3

# Documentación automática de API
drf-spectacular==0.27.0

# Base de datos (ya incluida en Python)
# sqlite3 viene con Python por defecto

# Dependencias adicionales recomendadas para desarrollo
# Pillow para manejo de imágenes (útil para campos ImageField)
Pillow>=10.4.0

# Para entornos de desarrollo
python-decouple==3.8

# Para manejo de variables de entorno (opcional pero recomendado)
python-dotenv==1.0.0

# Para pruebas (opcional)
pytest-django==4.5.2

# Para formateo de código (opcional)
black==23.7.0

# Para análisis estático de código (opcional)
flake8==6.0.0